refactor(mobile): simplify thread settings presentation
- Extract shared thread settings options and tests
- Simplify sheet closing and remove the unused bottom-sheet dependency
