+++ apps/mobile/src/features/threads/thread-settings-options.test.ts
@@ -1,0 +1,38 @@
+import type { ProviderOptionDescriptor } from "@t3tools/contracts";
+import { describe, expect, it } from "vite-plus/test";
+
+import { selectableChoices } from "./thread-settings-options";
+
+const effortDescriptor: Extract<ProviderOptionDescriptor, { type: "select" }> = {
+  id: "effort",
+  label: "Reasoning",
+  type: "select",
+  options: [
+    { id: "low", label: "Low" },
+    { id: "medium", label: "Medium", isDefault: true },
+    { id: "high", label: "High" },
+    { id: "ultrathink", label: "Ultrathink" },
+    { id: "ultracode", label: "Ultracode" },
+  ],
+  currentValue: "high",
+  promptInjectedValues: ["ultrathink"],
+};
+
+describe("selectableChoices", () => {
+  it("keeps ordinary provider choices in their declared order", () => {
+    expect(selectableChoices(effortDescriptor).map((choice) => choice.id)).toEqual([
+      "low",
+      "medium",
+      "high",
+    ]);
+  });
+
+  it("hides prompt-injected and workflow-trigger choices", () => {
+    expect(selectableChoices(effortDescriptor).map((choice) => choice.id)).not.toContain(
+      "ultrathink",
+    );
+    expect(selectableChoices(effortDescriptor).map((choice) => choice.id)).not.toContain(
+      "ultracode",
+    );
+  });
+});
